+++ src/controllers/index.ts
@@ -1,0 +1,11 @@
+import { AboutController } from 'src/controllers/about.controller';
+import { ApplyController } from 'src/controllers/apply.controller';
+import { AuthController } from 'src/controllers/auth.controller';
+import { WhitelistController } from 'src/controllers/whitelist.controller';
+
+export default [
+  AuthController,
+  AboutController,
+  ApplyController,
+  WhitelistController,
+];
