--- src/modules/index.ts
@@ -1,6 +1,0 @@
-import { AboutModule } from 'src/modules/about.module';
-import { WhitelistModule } from 'src/modules/whitelist.module';
-
-import { ApplyModule } from './apply.module';
-
-export default [WhitelistModule, ApplyModule, AboutModule];
